attmepted to smooth out the transition from the intro of the chapter.
Realized that "Example" is a new problem type, in particular, this problem type lends itself well to Steve's style of statment/question
writing.

The result of this is that either example needs to redefine "answer" or perhaps much better, we should make new answer types to accomodate
this statment style of writing. Whatever the answer type is, it will surely need two components, the "answer" and the "displayed" answer.

It could work something like this:

\fillin[$4$]{4}