Update diginWhatIsALimit.tex
When using a calculator to figure out the value of f(x) = sin(pi/x) at x of 0.1, 0.01, 0.001, and 0.0001, I generate the results 0.521, -0.717, -0.989, and 0.995 respectively. However, each of the questions have a correct answer of 0, which you get if you flip the fraction from sin(pi/x) to sin(x/pi). I'm using degrees on my calculator, which might be causing the confusion.

If it's not in the interest of the course to flip the fraction, it might be a good idea to mention somewhere that all calculations are in radians??

Loving the course so far, and the platform is amazing!