added many new exercises to the continuityOfFunctionsOfSeveralVariables section, split up some of the textbook.
I also made some modifications to the textbook sections in the continuityOfFunctionsOfSeveralVariables section, but left the original untouched.  Some changes include more examples, but others are changes I want for Math 1172 that may not be wanted for Math 2153, including making explicit that I want my students to think of points, not vectors, as inputs for functions of several variables.  In the exercises, I made sure that anything I added for the original text respects the convention taken there.

I am happy o discuss the content I added to the section with the Math 2153 team after the semester is over, or leave it at their discretion to adopt anything I did or discard it.