Ordered when the problems expand
One of the exercises was nested such that one of the problems was displayed: "Suppose that it is known that $\sum_{k=1}^{\infty} s_k$ diverges.  Consider the following examples: Thus, if we are only given that $\sum_{k=1}^{\infty} s_k$ diverges: ...".

I believe this to occur because both exercises have the same root exercise so they aren't appearing sequentially rather they are appearing one after another. It seemed like the conclusion of the problem was asked at the beginning.