Clarify description of exponents for polynomials
The exponents of x in polynomial functions are integers >= 0.

The used term "whole number" is ambiguous and can be misleading.
It can be interpreted as
- integer (including negative numbers)
or
- natural number
(see https://en.wikipedia.org/wiki/Integer)

According to https://en.wikipedia.org/wiki/Whole_number,
"whole number" is a colloquial term.

To be precise and to avoid misunderstandings, I suggest to use either
- non-negative integer
or
- natural number