Use better example for a polynomial as rational function.
Technically, f(x) = -4x^{-3}+5x^{-1}+7-18x^2 is a rational function
since it is the reduced version of
f(x) = \frac{-4+5x^{2}+7x^3-18x^5}{x^3}

However, that function in combination with the hint ("All polynomials
can be thought of as rational functions.") might be misleading.
A student might assume that f(x) = -4x^{-3} ... (with negative exponents)
is actually a polynomial (which is not the case).

I suggest to use a real polynomial as example.