wording of the problem
The wording of perpendicular to the x-axis was strange. I took out the first sentence:
We indicate the useful quantities on the base a prototypical slice:

and made the sentence after the graph be:
Since the cross section is a square, we let one of its sides lie on the base of the $xy$-plane, while the other extend vertically above the $xy$-plane.