Update understandingMathematicalLanguage.tex
Originally it was:
For the differential equation:
\[
3xy'+x^2y=y^3+3x
\]
the function $f(x,y,y')$ is $f(x,y,y') = 3xy'+\answer{x^2y-y^3-3x}$.

and function f is f= did not sound well so I changed it to we have the function, f= I changed this in two places that came up. Also I changed the wording of the sentence:
One advantage that this interpretation.... 
and closed the quotation marks.