Update thinkAboutItCenterOfSeries.tex
In the tables and afterwards pn and Pn have to be p2 and P2 so I fixed all of them. Some were pn and in the answer it was p3. 
Overall this problem has so many steps and hard to follow and the idea can easily get lost in it.