Update calculateATaylorPolynomialGuided.tex
In the sentence:
Now, note that we have exhibited several terms in the relevant series.  To find the sum of the first four nonzero terms in the sum, note that if we add the the expressions:

I took out one of "the"
Also later in the table for the two terms separately added together, they had (3ex)4cos(x) but it should be just 4cos(x) and afterwards when they add the terms from e^2x and 4cos(x) they get (1ex)e^2x +4cos(2x) but it should be e^(2x)+4cos(x) so I changed them. Also in a multiple choice question before it, it says the series e^(2x) +4cos(x) is valid up to and the answer it is up to x^4 term but this is confusing since these are valid up to any power.