Update powersOfTrigFunctions.tex
Changed:
When evaluating integrals of trigonometric functions with the same argument (i.e. inner function...the argument of $\sin(3x)$ would be $3x$, it is often....
To 
When evaluating integrals of trigonometric functions with the same argument (i.e. the inner function for $sin(3x)$ would be $3x$), it is often necessary...
So closed the parenthesis and changed the wording in the parenthesis.