Update l'HopitalsRuleReview.tex
Changed:
"The theorem above can be generalized to handle the case where $x \to \infty$ as well, and the results can be summarized by noting that when we have functions $f$ and $g$ that meet the criteria of the theorem

\begin{quote}
When a limit has the indeterminate form $\frac{0}{0}$ or $\pm \frac{\infty}{\infty}$, we can apply L'H\^{o}pital's Rule.
\end{quote}"

To:

\begin{quote}
The theorem above can be generalized to handle the case where $x \to \infty$ as well. Summarizing: if we have functions $f$ and $g$ that meet the criteria of the theorem and the limit has the indeterminate form $\frac{0}{0}$ or $\pm \frac{\infty}{\infty}$, then we can apply L'H\^{o}pital's Rule.
\end{quote}

The sentences were not complete and did not match before.