Update l'HopitalsRuleManipulation1.tex
Changed:
This example is one in which the indeterminate form $0 \times \pm \infty$ arise.  When this happens, we either must manipulate the original expression either to put

\begin{itemize}  
\item the expression whose limit is $0$ in the denominator.
\item the expression whose limit is $\pm \infty$ in the denominator.
\end{itemize}

To:

This example is one in which the indeterminate form $0 \times \pm \infty$ arise.  When this happens, we must manipulate the original expression by either putting 

\begin{itemize}  
\item the expression whose limit is $0$ in the denominator or
\item the expression whose limit is $\pm \infty$ in the denominator.
\end{itemize}