Update l'HopitalsRuleManipulation3.tex
Changed:
Direct evaluation gives the form $\infty - \infty$, which is an indeterminate form;We cannot use L'H\^{o}pital's Rule yet, but we can use the properties of logarithms.

To:

Direct evaluation gives the form $\infty - \infty$, which is an indeterminate form. We cannot use L'H\^{o}pital's Rule yet, but we can use the properties of logarithms.

So took out ; and made it a period.