Update trigonometricIntegrals.tex
In the sentence:
We studied trigonometric integrals that involve products of trigonometric functions with the same argument.  In these integrals, it was useful to have combinations of

I took out "with the same argument". 

Also I changed:

A helpful initial strategy is to try a substitution by setting $u$ equal to one of the trigonometric functions.  After saving a copy of the derivative for the substitution, if there is even number of the other trig function, we use the appropriate Pythagorean identity to convert it.

To:

A helpful initial strategy is to try a substitution by setting $u$ equal to one of the trigonometric functions.  After saving a copy of the derivative of $u$, if there is an even number of the other trig function, we use the appropriate Pythagorean identity to convert it to a function of $u$.