Update divergenceTestCompute2.tex
For the second question with bn=sin(1/n) the first and third choices were both coded as correct but the answer is only the third option which is the divergence test is inconclusive as it says after answering the question correctly. So I took out the code [correct] before the first choice.