Update ratioTestAndPolynomials.tex
After to gain a little practice the following equation is given:

\lim_{n \to \infty} \frac{p(n+1)}{p(n)} = \answer{\frac{(n+1)^5}{n^5}}

But the answer has n in it and there should be a limit before it so I put the limit before the answer as part of the question. This happens in the question after this and I fixed that as well.