Update 6_Q2-1151.tex
Problem url:
https://ximera.osu.edu/mooculus/asymptotesAsLimits/exercises/exerciseList/asymptotesAsLimits/exercises/6_Q2-1151

took out while at the same time, $x$ approaches $\answer{4}$ in:

As $x$ approaches $4$ from the right, $x-4$ is .... and approaching zero while at the same time, $x$ approaches ......

So they have to put in 4 at the end and it seems repetitive. I took out the ending:

As $x$ approaches $4$ from the right, $x-4$ is .... and approaching zero.