Update tangentLinesBothWays.tex
Problem url:
https://ximera.osu.edu/mooculus/parametricEquations/exercises/exerciseList/parametricEquations/exercises/tangentLinesBothWays

Fixed the sentence:
Thus the the $x$-coordinate and $y$-coordinate of any point on our given curve satisfies $y^{\answer{2}}=x^{\answer{3}}$.

(took out one of the "the" at the beginning)
Also put the hint in parenthesis in:

Notice that given our parametric equations that $x^3=\answer{t^6}$ and that $y^2=\answer{t^6}$ (give answers in terms of $t$).

Right after the answers so they see it. Before it was beneath the answer and is likely to be noticed after one answers the question.