Update tangentLine1.tex
Problem url:
https://ximera.osu.edu/mooculus/parametricEquations/exercises/exerciseList/parametricEquations/exercises/tangentLine1

Two sentences had the the and removed one of the ''the''
Also in:

We can find the horizontal tangent lines by examining when the derivative is $0$. That is when $dx/dt=0$.

have added "That is when $dx/dt=0$". because most books most books find them by setting dx/dt=0. In the problem this is not mentioned and they only say we find it by examining when the derivative becomes infinite. It is good to also mention this since it is easier to remember.