Update convertingPoints2.tex
https://ximera.osu.edu/mooculus/introductionToPolarCoordinates/exercises/exerciseList/introductionToPolarCoordinates/exercises/convertingPoints2

Added at the end:
(Give inverse trig functions as arcfunction than function^(-1). So arcsin(x) rather than sin^(-1)(x). Also don't give answers in decimals and type the expression involving the inverse trig function.)
This is the first problem that involves arctan and I want to make it clear how to put in the answers.