Update 14_Q4-1151_ex_1.tex
https://ximera.osu.edu/mooculus/higherOrderDerivativesAndGraphs/exercises/exerciseList/higherOrderDerivativesAndGraphs/exercises/14_Q4-1151_ex_1

Took out the first hint in:
\begin{hint}
Find the time, $t_{1}$, when the stone strikes the ground.
\end{hint}
\begin{hint}
Find the time, $t_{1}$, when the stone strikes the ground, by solving the equation
\[
s(t)=0
\]
and evaluate  $v(t_{1})$.
It was repetitive. Also put a hint:
Note that here we want speed not velocity 
so they note that it is +68 not -68.