Update velocityFromGraph2.tex
https://ximera.osu.edu/mooculus/higherOrderDerivativesAndGraphs/exercises/exerciseList/higherOrderDerivativesAndGraphs/exercises/velocityFromGraph2

This problem has a lot of questions and it might be better to split it up into two problems: one for velocity question and another for acceleration questions. Since the graph and the instructions on the graph kept on repeating and making it look even longer, I took out the graph and instructions after the first one. They can go on top of the page and see the graph for each question. This way it is more neat.