Update rectangleUnderCurve1.tex
https://ximera.osu.edu/mooculus/understandingFunctions/exercises/exerciseList/understandingFunctions/exercises/rectangleUnderCurve1

I changed x to b here. It called a random number x and it becomes difficult to solve because it gets confused with arbitrary x. If one uses b then y= 4-b^2 and setting the two ys equal to each other:
4-b^2 = 4-x^2/4 
one gets x=2b which makes the length of the rectangle 3b and then area is (3b)(4-b^2) which is the answer they have with x not b: