Update 05_03_ex_35.tex
https://ximera.osu.edu/mooculus/definiteIntegral/exercises/exerciseList/definiteIntegral/exercises/05_03_ex_35

Changed the wording of the first sentence. Also took out the hint at the end since they can do this directly by finding the integral and don't need geometry at this point. Some other problems were on integrals without the use of geometry so I think we have covered the material before this.