Update 06_01_ex_18.tex
https://ximera.osu.edu/mooculus/substitution/exercises/exerciseList/substitution/exercises/06_01_ex_18

Took out the hint. They have to be able to find what to put for u themselves. That is the main aspect of the substitution that they have to learn. I think the hint is giving out too much. The first few problems gave what to put for u and went through the process.