Update substitution1.tex
https://ximera.osu.edu/mooculus/substitution/exercises/exerciseList/substitution/exercises/substitution1

For the first question to get the upper bound for the integral, one is probably tempted to just look at the graph and say it is 1.25 but it is actually something they have to solve for by putting xcos(x^2) =0 so I added a hint saying find the x-intercept. 
Also later in the problem:
There is no g defined in the problem. It should be u(…) rather than g(…). So I fixed these.