Update relatedRatesApplied1.tex
https://ximera.osu.edu/mooculus/appliedRelatedRates/exercises/exerciseList/appliedRelatedRates/exercises/relatedRatesApplied1

Corrected a typo in the hint. Also please consider shortening or taking out some of the steps in the hint. It is giving away so much and there is nothing for someone to do for this problem. As the first problem in the assignment it might be good but still this is homework so they want to do problems themselves with hints not the process and the answer.