Update relatedRatesApplied2.tex
https://ximera.osu.edu/mooculus/appliedRelatedRates/exercises/exerciseList/appliedRelatedRates/exercises/relatedRatesApplied2

I fixed a typo at the beginning and also took out the hint. This is a simple circle problem and they should be able to do it without any help. The hint gives the whole process and the final answer. This is an exercise problem not a textbook example. So I think it is too much. The first problem in this assignment went through the process and gave the answer and I think that should be enough.