Update maxMinTrueFalse9.tex
https://ximera.osu.edu/mooculus/maximumsAndMinimums/exercises/exerciseList/maximumsAndMinimums/exercises/maxMinTrueFalse9

I think this can confuse students. We want them to associate decreasing with f' being negative so this statement makes them pause and doubt. So I added the sentence of what the conclusion should be to assure them that they can use the other direction of f' negative then f decreasing.